requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0

# LLM providers
openai>=1.3.0  # Optional: for OpenAI support
# Ollama uses HTTP API, no special package needed

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Utilities
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Web search
# google-search-results>=2.4.0  # SerpAPI (requires API key)
